Fixed reload issue with missing player data

+++ src/main/java/nl/imine/hubtweaks/PlayerDataManager.java
@@ -1,6 +1,7 @@
 package nl.imine.hubtweaks;
 
 import java.io.File;
+import java.util.UUID;
 
 import org.bukkit.Bukkit;
 import org.bukkit.World;
@@ -8,26 +9,24 @@ import org.bukkit.entity.Player;
 
 public class PlayerDataManager {
 
-	public static void RemoveAllPlayerData() {
+	public static void removeAllPlayerData() {
 		for (World w : Bukkit.getServer().getWorlds()) {
 			String worldDir = w.getName();
 			File file = new File(worldDir + File.separatorChar + "playerdata");
 			for (File f : file.listFiles()) {
-				System.out.println("Deleting: " + f.getName());
-				f.delete();
+				if (Bukkit.getPlayer(UUID.fromString(f.getName().split(".")[0])) != null) {
+					f.delete();
+				}
 			}
 		}
 	}
 
-	public static void RemovePlayerData(Player player) {
+	public static void removePlayerData(Player player) {
 		String UUID = player.getUniqueId().toString();
 		for (World w : Bukkit.getServer().getWorlds()) {
 			String worldDir = w.getName();
 			File file = new File(worldDir + File.separatorChar + "playerdata" + File.separatorChar + UUID + ".dat");
-			if (file.exists()) {
-				file.delete();
-				System.out.println("Deleting: " + file.getName());
-			}
+			file.delete();
 		}
 	}
 }
